Allow parameters for authentication.
Certain providers require additional parameters for actions during authentication.  For example Facebook will only show permissions once if they are declined.  The require that you include the auth_type parameter with a value of rerequest to show them again.  HybridAuth accounts for this by allowing parameters to be include in the second argument of authenticate.  This addition allows other programmers to use the filter to add their own parameters.  This should be done by adding the parameters to the existing array that is passed in.  The $provider value can be used to serve the proper parameters as needed.