Errors saying users table doesn't exist on Multisite Install
On a multisite install, I am getting errors that the users table doesn't exist. It is looking for the users table at wp_22_users using the prefix for the one site we are using it on. With a multisite install, the users table is typically wp_users and shared across the sites.

You should be able to fix that here: https://github.com/miled/wordpress-social-login/blob/master/includes/services/wsl.user.data.php#L48

By replacing {$wpdb->prefix}users with $wpdb->users which calls the users table.